excel_saver/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
XlsxWriter==3.2.2
yarl==1.18.3
